Adapted player model and renderer mixins from Player Animator
Adapted player model and renderer mixins from Player Animator that only worked for players, to now work for zombies. The goal is to allow zombies to animate their Better Combat attacks correctly. Currently, the game launches correctly and a test with a Husk shows no changes/errors when it attacks.