Adapted more player combat logic from Better Combat
Adapted combat logic from Better Combat that only worked for players, to now work for mpbs. The goal is to allow mobs to utilize the same attacking logic as players. Currently, the game launches correctly and a test with a Husk shows that it is unable to deal damage despite being in range and triggering an attack, and does not animate its attack. This means something in the chain from MeleeAttackGoal to MobAttackHelper#processAttack is not working correctly.