Attempted to debug animated attacks for Husks
Currently, the game launches correctly and a test with a Husk shows that it is unable to deal damage despite being in range and triggering an attack, and does not animate its attack. The updating of its combo count appears to be incorrect and attacks appear to only be animated for 2 ticks.