Implemented attack hitbox rendering for mobs
Added EntityRenderDispatcherMixin to render attack hitboxes for mobs. Added server-to-client syncing for a mob's combo count, so combo count can be used in client-side logic such as EntityRenderDispatcherMixin.renderColliderDebug. Fixed bug in MobTargetHelper.getRelation that would return TargetHelper.Relation.NEUTRAL for players, causing them to not be recognized as TargetHelper.Relation.HOSTILE for MobTargetFinder.getInitialTargets. Changed instances of passing in Mob.getTarget into calls of MobTargetFinder.findAttackTargets, since the attack target is not strictly guaranteed to be in the attack range of the mob. Fixed MobTargetFinder not using AttackRangeExtensions instead of MobAttackRangeExtensions.